Updated Employee Service

+++ src/main/java/com/vmware/poc/service/EmployeeService.java
@@ -13,46 +13,51 @@ import com.vmware.poc.repository.EmployeeRepository;
 public class EmployeeService {
 
 	private static final Logger logger = LoggerFactory.getLogger(EmployeeService.class);
-	private static final String EMPLOYEE_NOT_FOUND = "Employee not found with id:";
-	
+	private static final String NOT_FOUND = "Employee not found";
+
 	@Autowired
 	private EmployeeRepository employeeRepository;
 
 	public Employee saveAnEmployee(Employee employee) {
+		logger.debug("Saving an employee");
 		return employeeRepository.save(employee);
 	}
 
 	public Employee getAnEmployee(long id) {
 		Optional<Employee> employeeOptional = employeeRepository.findById(id);
-		if (employeeOptional.isPresent())
+		if (employeeOptional.isPresent()) {
+			logger.debug("Fetching an employee");
 			return employeeOptional.get();
-		logger.error(EMPLOYEE_NOT_FOUND + id);
+		}
+		logger.error(NOT_FOUND);
 		return null;
 	}
 
 	public Employee updateAnEmployee(long id, Employee employee) {
-
 		Employee storedEmployee = getAnEmployee(id);
 
 		if (storedEmployee == null) {
-			logger.error(EMPLOYEE_NOT_FOUND + id);
+			logger.error(NOT_FOUND);
 			return null;
 		}
 
-		if (!employee.getEmployeeName().isEmpty())
-			storedEmployee.setEmployeeName(employee.getEmployeeName());
-
-		if (employee.getEmployeeAge() > 0)
-			storedEmployee.setEmployeeAge(employee.getEmployeeAge());
-
+		logger.debug("Updating an employee");
 		return saveAnEmployee(storedEmployee);
 	}
 
-	public void deleteAnEmployee(long id) {
+	public boolean deleteAnEmployee(long id) {
+
+		if (getAnEmployee(id) == null)
+			return false;
+
+		logger.debug("Deleting employee");
 		employeeRepository.deleteById(id);
+		return true;
+
 	}
 
 	public List<Employee> getEmployeeList() {
+		logger.debug("Fetching all employees");
 		return (List<Employee>) employeeRepository.findAll();
 	}
 }
